Rename: 하트 아이콘 컴포넌트 폴더 구조 변경
